+++ core-ui/src/commonMain/kotlin/com/maksimowiczm/foodyou/core/ui/ext/LocalDateExt.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.saveable.Saver
 import kotlinx.datetime.LocalDate
 
 val LocalDate.Companion.Saver
-    get() = Saver<LocalDate, Int>(
+    get() = Saver<LocalDate, Long>(
         save = { it.toEpochDays() },
         restore = { fromEpochDays(it) }
     )
